+++ src/main/java/com/zepben/evolve/database/sqlite/tables/TableVersion.kt
@@ -14,7 +14,7 @@ import com.zepben.evolve.database.sqlite.tables.Column.Nullable.NOT_NULL
 @EverythingIsNonnullByDefault
 class TableVersion : SqliteTable() {
 
-    val SUPPORTED_VERSION = 27
+    val SUPPORTED_VERSION = 28
 
     val VERSION = Column(++columnIndex, "version", "TEXT", NOT_NULL)